% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_decorrelation.R
\name{adjustMatrixAsNeeded}
\alias{adjustMatrixAsNeeded}
\title{Whitens an input data matrix according to the covariance matrix input}
\usage{
adjustMatrixAsNeeded(X, covar, whitener = NULL)
}
\arguments{
\item{X}{\itemize{
\item N x M matrix of affects to "decorrelate"
}}

\item{covar}{\itemize{
\item covarnace matrix (M x M) with structure we are decorrelating for
}}

\item{whitener}{\itemize{
\item optional argument, an M x M matrix used to transform the data (from buildWhiteningMatrix)
}}
}
\value{
the decorrelated X
}
\description{
Whitens an input data matrix according to the covariance matrix input
}
