% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{BICglm}
\alias{BICglm}
\title{BIc calculation for use with glmnet}
\usage{
BICglm(fit, bic.mode = "")
}
\arguments{
\item{fit}{glmnet object}
}
\value{
BIC score for each call
}
\description{
BIc calculation for use with glmnet
}
