% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_decorrelation.R
\name{blockifyCovarianceMatrix}
\alias{blockifyCovarianceMatrix}
\title{Force a covariance matrix to be block diagonal
This matches our assumption that studies share covariance with a block diagonal structure (i.e. limited numbers are grouped)}
\usage{
blockifyCovarianceMatrix(blocks, covar)
}
\arguments{
\item{covar}{The actual covariance matrix; a matrix object}

\item{...}{}
}
\description{
Force a covariance matrix to be block diagonal
This matches our assumption that studies share covariance with a block diagonal structure (i.e. limited numbers are grouped)
}
