% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_obj.R
\name{calcGlobalResiduals}
\alias{calcGlobalResiduals}
\title{Helper function to get the residuals across all terms}
\usage{
calcGlobalResiduals(
  X,
  W,
  U,
  V,
  W_cov = NULL,
  fixed_first = FALSE,
  std.explanatory = FALSE,
  scalar = 1
)
}
\arguments{
\item{X}{Matrix of raw effect sizes (N x M)}

\item{W}{Matrix of weights corresponding to effect sizes}

\item{U}{Current estimate of loadings (U, N x K)}

\item{V}{Currest estimate of factors (V, M x K)}

\item{fixed_first}{Do we omit the first factor?}

\item{scalar}{is the fit scaled in any way?}

\item{W_c}{Inverse cholesky decomposition decorrelation matrix (M x M), optional}
}
\value{
the matrix of residuals, N x M
}
\description{
Helper function to get the residuals across all terms
}
