% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_optim.R
\name{CalcMatrixBIC.loglikGLOBALversion}
\alias{CalcMatrixBIC.loglikGLOBALversion}
\title{BIC using global fit calculation instead of sum of each regression's liklihood
Note that this doesn't require any fancy counting on the residuals, since all are just lumped into one regression}
\usage{
CalcMatrixBIC.loglikGLOBALversion(
  X,
  W,
  U,
  V,
  which.learning,
  W_cov = NULL,
  df = NULL,
  lm.fit.residuals = NULL,
  decomp = FALSE,
  ...
)
}
\arguments{
\item{X}{original data matrix (SNPs x studies)}

\item{W}{uncertainty associated with X (SNPs x studies)}

\item{U}{SNPs x K matrix}

\item{V}{Studies x K matrix}

\item{which.learning}{if we are learning "V" or "U"}

\item{W_cov}{Covariance matrix to correct for. For now just implemented in U.}

\item{df}{degrees of freedom associated with matrix currently learning}

\item{lm.fit.residuals}{Specify these to scale the log-liklihood by the optimal OLS fit. Not recommended.}
}
\description{
BIC using global fit calculation instead of sum of each regression's liklihood
Note that this doesn't require any fancy counting on the residuals, since all are just lumped into one regression
}
