% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{calculateBIC}
\alias{calculateBIC}
\title{Wrapper for all BIC methods under consideration.
TODO: In some instances, sklearn_ebic converts input matrices into "dense" matrix object. Want to avoid this.}
\usage{
calculateBIC(fit, covars, y, bic.mode)
}
\arguments{
\item{fit}{glmnet object}

\item{covars}{long.v or long.u}

\item{y}{long.x- the explanatory data (weighted)}

\item{bic.mode}{which version to use; options are Zou, dev, sklearn, or the extended variants of them.}
}
\value{
BIC score for each call (list);
Object is a list with bic.list (list of scores), fit.term (list of fit terms), df.term (degrees of freedom term), fit.scaler (denominator for fit term)
and addends (non-consant vector of additional terms, usually eBIC. Should not include constant single terms.), and n.coef- how many coefficients in the maximally testeed model.
}
\description{
Wrapper for all BIC methods under consideration.
TODO: In some instances, sklearn_ebic converts input matrices into "dense" matrix object. Want to avoid this.
}
