% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_selection_grid_search.R
\name{checkOverfitVar}
\alias{checkOverfitVar}
\title{Check if the current variance estimate is really off from all the others and give a warning.}
\usage{
checkOverfitVar(globalmin, remaining, fit_vect, min_index)
}
\arguments{
\item{remaining}{\itemize{
\item all other variance estimates
}}

\item{fit_vect}{\itemize{
\item fit data for all entries
}}

\item{min_index}{\itemize{
\item inde corresponding to the globally minimal variance estiamte
}}

\item{global}{\itemize{
\item globally minimial variance estimate
}}
}
\value{
nothing if selected variance estimate looks reasonable; otherwise it throws a warning
}
\description{
Check if the current variance estimate is really off from all the others and give a warning.
}
