% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_UV.R
\name{CheckVEmpty}
\alias{CheckVEmpty}
\title{Check if all the entries in a V matrix are 0 with a message}
\usage{
CheckVEmpty(V)
}
\arguments{
\item{V}{matrix to check}
}
\value{
Boolean if V is empty or not
}
\description{
Check if all the entries in a V matrix are 0 with a message
}
