% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_selection_grid_search.R
\name{chooseNextParams}
\alias{chooseNextParams}
\title{Figure out which points to test next
Currently limited to BINARY SEARCH- only searches 2 points next, at the midpoint between points that have been previously tested (integers only)}
\usage{
chooseNextParams(best_K, curr_grid, all_K)
}
\arguments{
\item{best_K-}{the K that is currently best}

\item{curr_grid-a}{list with all the BICs and all the Ks}

\item{all_K-}{all possible Ks to consider}
}
\value{
A list with the next K to test, and a logical saying if the search should end
}
\description{
Figure out which points to test next
Currently limited to BINARY SEARCH- only searches 2 points next, at the midpoint between points that have been previously tested (integers only)
}
