% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_obj.R
\name{computeObjIntermediate}
\alias{computeObjIntermediate}
\title{Wrapper on objective update, for in between each iteration- adds a check for alignment that shouldn't ever get triggered.}
\usage{
computeObjIntermediate(X, W, W_c, U, V, option, alpha, lambda, ...)
}
\arguments{
\item{X}{NxM matrix of SNP effect sizes}

\item{W}{NxM matrix of SNP weights (1/SE)}

\item{W_c}{MxM matrix for decorrelating transformation}

\item{U}{NxK matrix of SNP loadings}

\item{V}{MxK matrix of trait loadings}

\item{option}{list of gleanr options}

\item{alpha}{sparsity parameter corresponding to U}

\item{lambda}{sparsity parameter corresponding to V}

\item{...}{other arguments for compute_obj function}
}
\description{
Wrapper on objective update, for in between each iteration- adds a check for alignment that shouldn't ever get triggered.
}
