% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophenetic_calc.R
\name{copheneticAssessment}
\alias{copheneticAssessment}
\title{Calculate cophenetic correlation coefficient}
\usage{
copheneticAssessment(M)
}
\arguments{
\item{M}{a list of matrices}
}
\value{
coph a numerical correlation score in (0,1)
}
\description{
Calculate cophenetic correlation coefficient
}
