% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_UV.R
\name{cor2}
\alias{cor2}
\title{Faster correlation calculation for large matrices
Grabbed this off of stack overflow.}
\usage{
cor2(x)
}
\arguments{
\item{x}{}
}
\value{
Correlation value of a complex matrix.
}
\description{
Faster correlation calculation for large matrices
Grabbed this off of stack overflow.
}
