% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_UV.R
\name{DefineSparsitySpace}
\alias{DefineSparsitySpace}
\title{Title}
\usage{
DefineSparsitySpace(
  X,
  W,
  W_cov,
  fixed,
  learning,
  option,
  fit = "None",
  reg.elements = NULL
)
}
\arguments{
\item{X}{NxM matrix of SNP effect sizes}

\item{W}{NxM matrix of SNP weights (1/SE)}

\item{W_cov}{MxM matrix for decorrelating transformation}

\item{fixed}{the actual matrix object (either U or V) which is fixed}

\item{learning}{a string specifying which matrix is fixed (either "U" or "V")}

\item{option}{list of gleanr options}

\item{fit}{fit object from glmnet}
}
\description{
Title
}
