% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning_factors.R
\name{DropFactorsByObjective}
\alias{DropFactorsByObjective}
\title{A recursive function to drop factors that by so doing improve (that is, reduce) the overall objective}
\usage{
DropFactorsByObjective(X, W, W_c, U, V, maxK, option, minK = 0, scalar = 1)
}
\arguments{
\item{X}{full data matrix}

\item{W}{standard errors}

\item{W_c}{Whitening covariance matrix}

\item{U}{Predicted U}

\item{V}{Predicted V}

\item{maxK}{maximum K to include}

\item{option}{GLEANER options object}

\item{minK}{fewest number of factors to allow; default is 0}
}
\value{
updated U,V,K
}
\description{
A recursive function to drop factors that by so doing improve (that is, reduce) the overall objective
}
