% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_tools.R
\name{fillDefaultSettings}
\alias{fillDefaultSettings}
\title{Filling in settings, all relics of a past run.
TODO: basically get rid of all this}
\usage{
fillDefaultSettings(curr.args)
}
\arguments{
\item{curr.args}{}
}
\value{
updated arguments
}
\description{
Filling in settings, all relics of a past run.
TODO: basically get rid of all this
}
