% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_U.R
\name{FitUGlobal}
\alias{FitUGlobal}
\title{Performs fit step for of U all SNPs simultaneously}
\usage{
FitUGlobal(X, W, W_c, V, option, formerU, r.v = NULL, reg.elements = NULL)
}
\arguments{
\item{X}{unscaled coefficients for single SNP across M studies}

\item{W}{Standard errors associated with row of x SNPs}

\item{W_c}{the decorrelation covariance matrix used in the GLS form of this software.}

\item{V}{Fixed factor matrix from previous iteration}

\item{option}{List of setting options (here, function accesses: traitSpecificVar, gls, carry_coeffs,regression_method,actively_calibrating_sparsity)}

\item{formerU}{Matrix of U from previous iteration to use in this one}

\item{r.v}{List of residual variances- for use only when accounting for study-specific variance}

\item{reg.elements}{Pre-weighted regression elements}
}
\value{
a list containing: "l", the current learned row of U (k x 1) of weights for a given SNP, and the maximum possible sparsity for that run
}
\description{
Internal regression step for one SNP when learning U
Key options here include:
}
