% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_U.R
\name{FitUWrapper}
\alias{FitUWrapper}
\title{Wrapper for the U-fitting step}
\usage{
FitUWrapper(X, W, W_c, V, option, prevU = NULL, ...)
}
\arguments{
\item{X}{the matrix to be decomposed (N x T, N is the number of data points, T is the number of features)}

\item{W}{the weight matrix, same size as in X, containing standard errors of the GWAS estimates}

\item{W_c}{the whitening matrix, associated with covariance structure, a M x M}

\item{V}{learned factor matrix  (M x K, M is the number of features, K is the number of factors)}

\item{option}{list with parameters including lambda1, the l1 penalty parameter for the U matrix
(relevant arguments include: ncores,)}

\item{...}{arguments passed on to FitU}
}
\value{
A list containing the return matrix U, the distribution of maximum sparsity parameters, and any columns dropped because they contained NAs
}
\description{
Wrapper for the U-fitting step
}
