% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_V.R
\name{FitVGlobal}
\alias{FitVGlobal}
\title{Update the factor matrix V during the alternating least squares}
\usage{
FitVGlobal(X, W, W_c, U, option, formerV = NULL, reg.elements = NULL)
}
\arguments{
\item{X}{NxM matrix of SNP effect sizes}

\item{W}{NxM matrix of SNP weightings (1/SE)}

\item{W_c}{MxM decorrelating transformation matrix}

\item{U}{NxK matrix of curent estimated SNP loadings}

\item{option}{list of options for gleanr}

\item{formerV}{\itemize{
\item option to submit previous iteration's V. Not used
}}

\item{reg.elements}{\itemize{
\item list with pre-transformed outcome variable and weighting data. Helps with memory and speed
}}
}
\value{
list containing updated V and several additional key statistics, including BIC information
}
\description{
Update the factor matrix V during the alternating least squares
}
