% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_selection_grid_search.R
\name{gatherSearchData}
\alias{gatherSearchData}
\title{Manages results of grid search and nominates next parameters to look at}
\usage{
gatherSearchData(
  gs_object,
  k_range,
  grid_record,
  curr_grid = NULL,
  curr_best = NULL,
  version = "global"
)
}
\arguments{
\item{gs_object}{\itemize{
\item output of grid_search function
}}

\item{k_range}{\itemize{
\item full range of possible K to consider searching
}}

\item{...}{May be curr_grid=sdv$query_matrix, the current matrix of searches, or
curr_best = sdv$query_matrix$min_result, the best bic_dat object so far}
}
\value{
a list containing
}
\description{
Manages results of grid search and nominates next parameters to look at
}
\examples{
#gatherSearchData(next.grid,k.range, curr_grid=sdv$query_matrix,curr_best = sdv$min_result )
}
