% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_selection_grid_search.R
\name{getNewTestBounds}
\alias{getNewTestBounds}
\title{Helper function for chooseNextParams to get the bounds of the space to test next.}
\usage{
getNewTestBounds(best_K, curr_grid, all_K)
}
\arguments{
\item{best_K}{K setting which minimzes the BIC}

\item{curr_grid}{list of previously evaluated K points}

\item{all_K}{range of valid K for consideration}
}
\value{
upper (above) and lower (below)
}
\description{
Helper function for chooseNextParams to get the bounds of the space to test next.
}
