% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_obj.R
\name{GetStepWiseObjective}
\alias{GetStepWiseObjective}
\title{Calculate the change in objective with each individual matrix update
DEPRECATED- this doesn't make sense anymore, since we aren't updating one column or row at a time}
\usage{
GetStepWiseObjective(
  X,
  W,
  W_c,
  old.mat,
  new.mat,
  companion.mat,
  fixed.term,
  option
)
}
\arguments{
\item{X}{data}

\item{W}{uncertainty}

\item{W_c}{decorrelation matrix}

\item{old.mat}{the old version of the matrix you just updated}

\item{new.mat}{the new version of the matrix you just updated}

\item{companion.mat}{the complement of the matrix you just updated}

\item{fixed.term}{which term the complement is (u or v)}

\item{option}{list of options}
}
\value{
at list of updated objectives with each change
}
\description{
Calculate the change in objective with each individual matrix update
DEPRECATED- this doesn't make sense anymore, since we aren't updating one column or row at a time
}
