% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{gleanr}
\alias{gleanr}
\title{Wrapper to programmatically call the BIC-based factorization}
\usage{
gleanr(
  X,
  W,
  snp.ids,
  trait.names,
  C = NULL,
  K = "GRID",
  covar_se = NULL,
  bic.var = "sklearn_eBIC",
  use.init.k = FALSE,
  init.mat = "V",
  is.sim = FALSE,
  save.path = "",
  scale.mats = FALSE,
  regression.method = "glmnet",
  save_out = TRUE,
  shrinkWL = -1,
  ...
)
}
\arguments{
\item{X}{NxM matrix of SNP effect sizes}

\item{W}{uncertainty weights (1/SE), same dimensions as X}

\item{snp.ids}{the names of the SNPs in X, in order}

\item{trait.names}{the names of the studies in X, in order}

\item{K}{the target number of K to yield.}

\item{bic.var}{How to calculate the variance on BIC term. Deprecated as of Feb 21, 2023}

\item{use.init.k}{To initialize to given K or to full}

\item{init.mat}{Which matrix to initialize with, V or U. Default is V}

\item{is.sim}{Specify if this run is a simulation. Default is false}

\item{save.path}{save results along the way as output (optional)}
}
\description{
Wrapper to programmatically call the BIC-based factorization
}
