% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_selection_grid_search.R
\name{gridSearchK}
\alias{gridSearchK}
\title{Find the K that minimize the BIC via grid search}
\usage{
gridSearchK(
  opath,
  option,
  X,
  W,
  W_c,
  all_ids,
  names,
  step.limit = 8,
  init.range = 4,
  ...
)
}
\arguments{
\item{opath}{to write out the K grid search data object}

\item{option}{to specify gleaner running settings}

\item{X}{NxM matrix of SNP effect sizes}

\item{W}{NxM matrix of SNP weights}

\item{W_c}{MxM matrix decorrelating transform}

\item{all_ids}{SNP ids corresponding to X}

\item{names}{Trait names corresponding to X}

\item{step.limit}{\itemize{
\item what is the maximum number of grid look steps you are willing to take?
}}

\item{init.range}{\itemize{
\item how broad should the initial search range be? Default is 4
}}

\item{...}{}

\item{parallel}{\itemize{
\item options are "none" or "multicore" or "snow"
}}

\item{ncpu}{\itemize{
\item "how many CPUs to use if going parallel
}}
}
\value{
the bic_dat object with the best BIC
}
\description{
Find the K that minimize the BIC via grid search
}
