% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{gwasML_ALS_Routine}
\alias{gwasML_ALS_Routine}
\title{Model fitting routine, using the updateUV function}
\usage{
gwasML_ALS_Routine(
  option,
  X,
  W,
  W_c,
  optimal.init,
  maxK = 0,
  opath = "",
  no.prune = FALSE,
  reg.elements = NULL,
  ...
)
}
\arguments{
\item{option}{list of runtime options}

\item{X}{NxM matrix of SNP effect sizes}

\item{optimal.init}{V at which to initialize the run, typically the output from a model selection step}

\item{maxK}{specify how many K max to include. Factors will be removed if they exceed this.}

\item{opath}{optional path to write out to for saving RData}

\item{no.prune}{if no pruning of factors is to occur}

\item{reg.elements}{list with pre-weighted regression variables}

\item{...}{}
}
\value{
list object with final U and V and convergence information
}
\description{
Model fitting routine, using the updateUV function
}
