% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{initializeGLEANR}
\alias{initializeGLEANR}
\title{Utilities for loading GLEANR by function, as opposed to the command line tools for reading everything in}
\usage{
initializeGLEANR(
  X,
  W,
  C,
  snp.ids,
  trait.names,
  K = 0,
  init.mat = "V",
  covar_shrinkage = -1,
  enforce_blocks = TRUE,
  covar_se = NULL,
  is.sim = FALSE,
  ...
)
}
\arguments{
\item{X}{NxM matrix of SNP effect sizes}

\item{C}{\itemize{
\item cohort overlap estimates
}}

\item{snp.ids}{\itemize{
\item list of SNP ids, corresponds to order in X∂
}}

\item{trait.names}{\itemize{
\item trait names to use, corresponds to order in X
}}

\item{K}{\itemize{
\item specify a K if you would like.
}}

\item{covar_shrinkage}{Degree of shrinkage towards the identity. matrix to apply to C (0 means none, 1 means perfect shrhinkage.) If covar_se is not NULL, defaults to the variance based (Strimmer) approach}

\item{enforce_blocks}{Force the matrix to be block matrix, default TRUE}

\item{covar_se}{Matrix of standard errors corresponding to C. Used for shrinkage. If given, covar_shrinkage defaults to "STRIMMER"}

\item{...}{}

\item{W-}{weights from Standard errors}

\item{init.mat-}{which matrix to initialize to}
}
\value{
a list containing the objects needed to run gleaner, including the decorrelating transformation matrix andthe options matrix.
}
\description{
Utilities for loading GLEANR by function, as opposed to the command line tools for reading everything in
}
