% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_decorrelation.R
\name{linearShrinkLWSimple}
\alias{linearShrinkLWSimple}
\title{Simplifed WL shrinkage using a}
\usage{
linearShrinkLWSimple(sample_cov_mat, gamma)
}
\arguments{
\item{sample_cov_mat}{the input covariance matrix used, from LDSC}

\item{gamma}{the setting at which to run this}
}
\value{
shrunk estimate
}
\description{
Simplifed WL shrinkage using a
}
