% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{MatrixDFU}
\alias{MatrixDFU}
\title{Estimator of the lasso degrees of freedom as the # of non-zero coefficients}
\usage{
MatrixDFU(mat_in, fixed_first = FALSE)
}
\arguments{
\item{mat_in}{to get DF of}

\item{fixed_first}{if we count the first factor (applies to V only)}
}
\description{
Estimator of the lasso degrees of freedom as the # of non-zero coefficients
}
