% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_tools.R
\name{matrixGWASQC}
\alias{matrixGWASQC}
\title{Clean up input GWAS data, including extreme X^2 values and entries with too many NAs
TODO: add inheritance for some of these arguments}
\usage{
matrixGWASQC(X, W, id.list, na_thres = 0.5, na.threshold = 0.7)
}
\arguments{
\item{na.threshold}{}
}
\value{
a list of updated input items (X,W,id.list) reflecting the drops/filtering
}
\description{
Clean up input GWAS data, including extreme X^2 values and entries with too many NAs
TODO: add inheritance for some of these arguments
}
