% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsity_scaler.R
\name{matrixSparsity}
\alias{matrixSparsity}
\title{Calculate the proportion of empty cells in a matrix}
\usage{
matrixSparsity(m, initK, thresh = 0, wrt.init = FALSE)
}
\arguments{
\item{m}{matrix to estimate sparsity of}

\item{initK}{inital K, if the matrix is a sub-matrix (e.g. via dropped columns)}

\item{thresh}{threshold for counting sparsity}

\item{wrt.init}{if you want sparsity with respect to full inintal size of matrix (e.g. N x K)}
}
\value{
Proportion of cells with absolute values below threshold
}
\description{
Calculate the proportion of empty cells in a matrix
}
