% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pve.R
\name{mean.na}
\alias{mean.na}
\title{Helper function to omit NAs from mean calculation vector.
Review,this can be replaced with mean(..., na.rm=TRUE)}
\usage{
\method{mean}{na}(vec)
}
\arguments{
\item{vec}{}
}
\value{
mean of vector sans NAs
}
\description{
Helper function to omit NAs from mean calculation vector.
Review,this can be replaced with mean(..., na.rm=TRUE)
}
