% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_tools.R
\name{orderColumnsByName}
\alias{orderColumnsByName}
\title{Function to order the columns by matching phenotype names
.}
\usage{
orderColumnsByName(query.mat, ref, force.ref = "")
}
\arguments{
\item{query.mat}{the matrix to check the order on}

\item{ref}{the reference order of phenotypes}

\item{force.ref}{the matrix doesn't yet have names, so just force the names in the order given in ref}
}
\value{
an ordered matrix
}
\description{
Function to order the columns by matching phenotype names
.
}
