% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning_factors.R
\name{OrderEverythingByPVE}
\alias{OrderEverythingByPVE}
\title{Order a return object so factors in order by PVE}
\usage{
OrderEverythingByPVE(ret, X, W, W_c, options, terminal.attempt = FALSE, ...)
}
\arguments{
\item{ret}{the object return from ALS, with a V, U and PVE object ($)}

\item{X, W}{to calculate the new PVE (its possible V,U have changed..)}
}
\value{
return.dat, ret object updated by order of PVE
}
\description{
Order a return object so factors in order by PVE
}
