% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_obj.R
\name{penLL}
\alias{penLL}
\title{Calculate the log-liklihood of a data fit as given by penalized}
\usage{
penLL(n, resids)
}
\arguments{
\item{n}{the number of samples involved}

\item{resids}{the residuals involved, in a list}
}
\value{
the log-liklihood
}
\description{
Calculate the log-liklihood of a data fit as given by penalized
}
