% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_obj.R
\name{penLLEmpirical}
\alias{penLLEmpirical}
\title{Calculate the log-liklihood of the data fit as actually optimized by penalized
This is the log-normal distribution, where we assume variance =1}
\usage{
penLLEmpirical(n, resids)
}
\arguments{
\item{n}{the number of samples involved}

\item{resids}{the residuals involved, in a list}
}
\value{
the log-liklihood
}
\description{
Calculate the log-liklihood of the data fit as actually optimized by penalized
This is the log-normal distribution, where we assume variance =1
}
