% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_UV.R
\name{prepRegressionElements}
\alias{prepRegressionElements}
\title{Generate regression explanatory variable and weighted elements}
\usage{
prepRegressionElements(X, W, W_c, option)
}
\arguments{
\item{X}{NxM matrix of SNP effect sizes}

\item{W}{NxM matrix of SNP weights (1/SE)}

\item{W_c}{MxM matrix for decorrelating transformation}
}
\value{
list containing the weight matrices (the per-snp weights, used in estimating V) and the weighted x Matrix
}
\description{
Generate regression explanatory variable and weighted elements
}
