% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning_factors.R
\name{PruneNumberOfFactors}
\alias{PruneNumberOfFactors}
\title{Cut out factors based on objective or some threshold (maxK)}
\usage{
PruneNumberOfFactors(X, W, W_c, reg.run, minK, maxK, option)
}
\arguments{
\item{X}{full effect size data}

\item{W}{effect size uncertainty}

\item{W_c}{covariance structure of effect sizes}

\item{reg.run}{output from update_UV; full ALS output}

\item{minK}{Minimum allowed # of factors. NOT USED}

\item{maxK}{Maximum allowecx # of factors}

\item{option}{}
}
\value{
updated ALS output list (same as update_UV output)
}
\description{
Cut out factors based on objective or some threshold (maxK)
}
