% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_tools.R
\name{readInData}
\alias{readInData}
\title{Load the relevant datasets for analysis based on an argument vector}
\usage{
readInData(args)
}
\arguments{
\item{args}{Specifies the paths to files and program settings to run. Includes \emph{gwas_effects, weighting_scheme, uncertainty, genomic_correction,} and \emph{covar_matrix arguments}}
}
\value{
A list with entries:
"X": sorted GWAS effect sizes,
"W": sorted GWAS uncertainty weights (1/SE),
"ids": sorted ID order, "trait_names" = names of the traits,
"C": the read in covariance matrix and "W_c": the (blockified) matrix to be used for decorrelation
}
\description{
Load the relevant datasets for analysis based on an argument vector
}
