% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_tools.R
\name{readInSettings}
\alias{readInSettings}
\title{Take command-line arguments and convert them into the gleanr options object}
\usage{
readInSettings(args)
}
\arguments{
\item{args, }{a list as you would get from a command line argument parser}
}
\value{
an options list object
}
\description{
Take command-line arguments and convert them into the gleanr options object
}
