% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_and_output.R
\name{reportSimilarities}
\alias{reportSimilarities}
\title{Debugging tool to print output as tracking progress}
\usage{
reportSimilarities(f, s, t)
}
\arguments{
\item{f}{vector of sparsity parameters}

\item{s}{vector of sparsity parameters}

\item{t}{vector of sparsity parameters}
}
\value{
none, just prints out
}
\description{
Debugging tool to print output as tracking progress
}
