% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{returnCompletionDat}
\alias{returnCompletionDat}
\title{Wrapper for returning program data}
\usage{
returnCompletionDat(min.dat, rec.dat, mat.fit, conv.options, general.options)
}
\arguments{
\item{min.dat}{tracker for minimal/optimal data}

\item{rec.dat}{record data from iteration to iteratino}

\item{mat.fit}{u.fit data object}

\item{conv.options}{record of convergence data}

\item{general.options}{list of program options.}
}
\value{
list with optimal U and V, optimal lambda and alpha, and some other helpful stuff.
}
\description{
Wrapper for returning program data
}
