% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_tools.R
\name{SampleOverlapCovarHandler}
\alias{SampleOverlapCovarHandler}
\title{Wrapper to get all the important data extracted and used for cohort overlap covariance adjjustment}
\usage{
SampleOverlapCovarHandler(args, names, X)
}
\arguments{
\item{args}{argument vector from initializing or calling gleanr.}

\item{names}{list of trait names, in order corresponding to X}

\item{X}{NxM matrix of SNP effect sizes}
}
\value{
list containing Whitening matrix W_c inverse, Block versino of the C matrix and unblocked version of C matrix.
}
\description{
Wrapper to get all the important data extracted and used for cohort overlap covariance adjjustment
}
