% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_tools.R
\name{selectInitK}
\alias{selectInitK}
\title{Select the K to initialize with. A few different options, made for testing.}
\usage{
selectInitK(args, X_, evals = NULL)
}
\arguments{
\item{args}{a list with settings; key settings are K desired (either a number or a method)}

\item{X_}{matrix, weighted and adjusted for covariance}

\item{evals}{(optional) the egienvalues to evaluate}
}
\value{
a number of K to initialize with
}
\description{
Select the K to initialize with. A few different options, made for testing.
}
