% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{sklearnBIC}
\alias{sklearnBIC}
\title{Calculate BIC based on what is in the sci-kit learn package (see https://scikit-learn.org/stable/modules/linear_model.html#lasso-lars-ic)
this differs from the traditional derivation, in that the use the unbiased estimate for s rather than the MLE.
Source code is here https://github.com/scikit-learn/scikit-learn/blob/872124551/sklearn/linear_model/_least_angle.py#L1991}
\usage{
sklearnBIC(fit, x, y)
}
\arguments{
\item{fit}{glmnet fit object of current regression run across many sparsity settings}

\item{x}{expanded sparse block matrix corresponding to U or V used in regression by FitVGlobal or FitUGlobal}

\item{y}{long vector of all stacked, weighted, and transformed SNP effect sizes}
}
\description{
Calculate BIC based on what is in the sci-kit learn package (see https://scikit-learn.org/stable/modules/linear_model.html#lasso-lars-ic)
this differs from the traditional derivation, in that the use the unbiased estimate for s rather than the MLE.
Source code is here https://github.com/scikit-learn/scikit-learn/blob/872124551/sklearn/linear_model/_least_angle.py#L1991
}
