% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{stdBIC}
\alias{stdBIC}
\title{Calculate the BIC using a vanilla, textbook standard formulation from a GLMNET fit object}
\usage{
stdBIC(fit, long.v, long.x)
}
\arguments{
\item{fit}{the GLMNET fit object}

\item{long.v}{the covariates matrix, sparse}

\item{long.x}{the outcome variable}
}
\value{
a vector of BIC scores across multiple different lambda settings
}
\description{
Calculate the BIC using a vanilla, textbook standard formulation from a GLMNET fit object
}
