% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_selection_grid_search.R
\name{storeGridResults}
\alias{storeGridResults}
\title{Update search grid with newest search data.}
\usage{
storeGridResults(gs_object, curr_grid = NULL)
}
\arguments{
\item{gs_object}{list of bic_autofit objects}

\item{curr_grid}{current data containing a list of Kinits and global BIC scores}
}
\value{
an updated grid with the latest looks.
}
\description{
Update search grid with newest search data.
}
