% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_decorrelation.R
\name{strimmerCovShrinkage}
\alias{strimmerCovShrinkage}
\title{Implementation of the covarinage shrinkage proposed by Schaefer and Strimmer in 2005,
"A shrinkage approach to large-scale covariance matrix estimation and implications for functional genomics"
(2005 "Statistical Applications in Genetics and Molecular Biology")}
\usage{
strimmerCovShrinkage(args, covar, covar_se, sd.scaling = 1)
}
\arguments{
\item{covar-}{covariance matrix, M x M}

\item{covar_se-}{SE for covariance estimates, M x M}

\item{sd.scaling-}{Scaling matrix based on standard deviation of Z-scores}
}
\description{
Implementation of the covarinage shrinkage proposed by Schaefer and Strimmer in 2005,
"A shrinkage approach to large-scale covariance matrix estimation and implications for functional genomics"
(2005 "Statistical Applications in Genetics and Molecular Biology")
}
