% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{trackAlternativeConvergences}
\alias{trackAlternativeConvergences}
\title{Track alternative source of convergence. Primarily experimental, to identify alternatives to the objective}
\usage{
trackAlternativeConvergences(
  X,
  W,
  W_c,
  U,
  V,
  index,
  record.data,
  u.alphamax,
  v.lambdamax,
  option_old
)
}
\arguments{
\item{X}{NxM matrix of SNP effect sizes}

\item{W}{NxM matrix of SNP weights (1/SE)}

\item{W_c}{MxM matrix for decorrelating transformation}

\item{U}{NxK matrix of SNP loadings}

\item{V}{MxK matrix of trait loadings}

\item{index}{whcih iteration we are currently on}

\item{v.lambdamax}{the maximum possible sparsity parameter for a given input X, U (||y'X||_INF)}

\item{option_old}{current option settings}

\item{u.alphamaxthe}{maximum possible sparsity parameter for a given input X, V (||y'X||_INF)}
}
\value{
track.modes list containing possible convergence criteria, including V and U norms, the objective function, BIC scores, their sum, and the U and V matrices
}
\description{
Track alternative source of convergence. Primarily experimental, to identify alternatives to the objective
}
