% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_decorrelation.R
\name{truncatedCovarMatrix}
\alias{truncatedCovarMatrix}
\title{Quickly estimate the correlation matrix that comes from non-genetic effects
Baesd on the method described here by Sebanti:https://deepblue.lib.umich.edu/bitstream/handle/2027.42/143992/sebanti_1.pdf?sequence=1}
\usage{
truncatedCovarMatrix(Z, z_thresh = 2)
}
\arguments{
\item{Z}{matrix of Z scores with dimensions n x m}

\item{z_thresh}{threshold of Z scores to omit, default is 2}
}
\value{
an
}
\description{
Quickly estimate the correlation matrix that comes from non-genetic effects
Baesd on the method described here by Sebanti:https://deepblue.lib.umich.edu/bitstream/handle/2027.42/143992/sebanti_1.pdf?sequence=1
}
