% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_UV.R
\name{update_UV}
\alias{update_UV}
\title{Main ALS workhorse with fixed sparsity parameters
Proceeds until some convergence threshold}
\usage{
update_UV(
  X,
  W,
  W_c,
  option,
  preV = NULL,
  preU = NULL,
  burn.in = FALSE,
  reg.elements = NULL
)
}
\arguments{
\item{X}{NxM matrix of SNP effect sizes}

\item{W}{NxM matrix of SNP weights (1/SE)}

\item{W_c}{MxM matrix for decorrelating transformation}

\item{option}{\itemize{
\item specifies major settings
}}

\item{preV}{\itemize{
\item V to start from
}}

\item{preU}{\itemize{
\item U to start from, if desired (option$u_init must be TRUE)
}}

\item{burn.in}{\itemize{
\item If you want to have some burn in iterations using un-regularized least squares.
}}

\item{reg.elements}{\itemize{
\item list of pre-weighted regression data.
}}
}
\description{
Main ALS workhorse with fixed sparsity parameters
Proceeds until some convergence threshold
}
