% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{UpdateAndCheckSparsityParam}
\alias{UpdateAndCheckSparsityParam}
\title{Update sparsity parameter list with next selection
Notify user if there was a large jump in the sparsity parameters.}
\usage{
UpdateAndCheckSparsityParam(prev.list, new, errorReport = FALSE)
}
\arguments{
\item{prev.list}{list of sparsity parameters so far}

\item{new}{new parameter proposed}

\item{errorReport}{Setting if we want to give extra information}
}
\value{
updated list of sparsity parameters
}
\description{
Update sparsity parameter list with next selection
Notify user if there was a large jump in the sparsity parameters.
}
