% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_and_output.R
\name{updateLog}
\alias{updateLog}
\title{Report update in job and write to log}
\usage{
updateLog(text, options = NULL)
}
\arguments{
\item{text}{What to write out}

\item{options}{Option list. Key values are 'ofile', which specifies where to write the log (or not to write to it at all if NULL),
and "V", which indcates how verbose to me}
}
\value{
None
}
\description{
Report update in job and write to log
}
