% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{updateRecDat}
\alias{updateRecDat}
\title{Helper function to update record data showing sparsity and bic changes}
\usage{
updateRecDat(
  rec.dat,
  mat.fit,
  matrix_on,
  iter,
  X,
  W,
  W_c,
  optimal_complement,
  option
)
}
\arguments{
\item{rec.dat}{current record data, a list of lists}

\item{mat.fit}{the fit of the most recent step, either u.fit or v.fit}

\item{matrix_on}{which matrix you just updated, either "U" or "V"}

\item{iter}{which iteration you're on}

\item{X}{NxM matrix of SNP effect sizes}

\item{W}{NxM matrix of SNP weights (1/SE)}

\item{W_c}{MxM matrix for decorrelating transformation}

\item{optimal_complement}{the opposite matrix of matrix_on (actual matrix object)}

\item{option}{}
}
\value{
updated rec.dat
}
\description{
Helper function to update record data showing sparsity and bic changes
}
