% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_tools.R
\name{writeRunReport}
\alias{writeRunReport}
\title{Quick readout to report settings, useful for debugging}
\usage{
writeRunReport(argsin)
}
\arguments{
\item{argsin}{the current arguments in.}
}
\description{
Quick readout to report settings, useful for debugging
}
