% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{initializeGLEANR}
\alias{initializeGLEANR}
\title{Utilities for loading GLEANR by function, as opposed to the command line tools for reading everything in}
\usage{
initializeGLEANR(
  X,
  W,
  C,
  snp.ids,
  trait.names,
  K = 0,
  init.mat = "V",
  covar_shrinkage = -1,
  enforce_blocks = TRUE,
  covar_se = NULL,
  ...
)
}
\arguments{
\item{X}{NxM matrix of SNP effect sizes}

\item{C}{\itemize{
\item cohort overlap estimates
}}

\item{snp.ids}{\itemize{
\item list of SNP ids, corresponds to order in X
}}

\item{trait.names}{\itemize{
\item traitnames to use, corresponds to order in X
}}

\item{K}{\itemize{
\item specify a K if you would like.
}}

\item{enforce_blocks}{\itemize{
\item Force the matrix to be block matrix
}}

\item{...}{}

\item{W-}{weights from Standard errors}

\item{init.mat-}{which matrix to initialize to}

\item{covar_shrinkage-}{Which covariance matrix shrinkage type to apply, default is variance based (Strimmer)}
}
\value{
a list containing the objects needed to run gleaner, including the decorrelating transformation matrix andthe options matrix.
}
\description{
Utilities for loading GLEANR by function, as opposed to the command line tools for reading everything in
}
